backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
python-dotenv==1.0.1
whitenoise==6.7.0
psycopg2-binary==2.9.9
gunicorn==21.2.0

